MSW 폴더 구조 변경 및 mockResponse 유틸 response 생성 (#940)
