+++ frontend/jest.config.ts
@@ -12,7 +12,7 @@ const config: Config = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
-      '<rootDir>/src/mocks/fileTransformer.js',
+      '<rootDir>/src/mocks/settings/fileTransformer.js',
   },
   clearMocks: true,
   moduleNameMapper: {
