+++ frontend/src/mocks/fixtures/categoryList.json
