+++ frontend/src/mocks/settings/browser.ts
