+++ frontend/src/queries/categories/useCategoryDeleteMutation.ts
@@ -1,9 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { QUERY_KEY, deleteCategory } from '@/api';
+import { ApiError } from '@/api/Error';
 import { ToastContext } from '@/contexts';
 import { useCustomContext } from '@/hooks';
-import { Category, CustomError } from '@/types';
+import { Category } from '@/types';
 
 export const useCategoryDeleteMutation = (categories: Category[]) => {
   const queryClient = useQueryClient();
@@ -14,14 +15,17 @@ export const useCategoryDeleteMutation = (categories: Category[]) => {
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [QUERY_KEY.CATEGORY_LIST] });
     },
-    onError: (error: CustomError, targetCategory) => {
-      const categoryId = targetCategory.id;
-      const categoryName = categories.find((category) => category.id === categoryId)?.name || '카테고리를 찾을 수 없음';
+    onError: (error, targetCategory) => {
+      if (error instanceof ApiError) {
+        const categoryId = targetCategory.id;
+        const categoryName =
+          categories.find((category) => category.id === categoryId)?.name || '카테고리를 찾을 수 없음';
 
-      if (error.status === 400) {
-        failAlert(`템플릿이 존재하는 카테고리(${categoryName})는 삭제할 수 없습니다.`);
-      } else {
-        failAlert(`카테고리 삭제 중 오류가 발생했습니다: ${categoryName}`);
+        if (error.statusCode === 400) {
+          failAlert(`템플릿이 존재하는 카테고리(${categoryName})는 삭제할 수 없습니다.`);
+        } else {
+          failAlert(`카테고리 삭제 중 오류가 발생했습니다: ${categoryName}`);
+        }
       }
     },
   });
