+++ frontend/src/queries/categories/useCategoryListQuery.spec.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 
 import { AuthProvider } from '@/contexts';
-import { categories as mockCategories } from '@/mocks/categoryList.json';
+import { categories } from '@/mocks/fixtures/categoryList.json';
 
 import { useCategoryListQuery } from './useCategoryListQuery';
 
@@ -14,13 +14,13 @@ const queryWrapper = ({ children }: { children: React.ReactNode }) => (
   </QueryClientProvider>
 );
 
-describe('useTagListQuery', () => {
+describe('useCategoryListQuery', () => {
   it('카테고리 목록을 id 오름차순으로 조회할 수 있다.', async () => {
     const { result } = renderHook(() => useCategoryListQuery({ memberId: 1 }), { wrapper: queryWrapper });
 
     await waitFor(() => {
       expect(result.current.isSuccess).toBe(true);
-      expect(result.current.data?.categories).toEqual(mockCategories);
+      expect(result.current.data.categories).toEqual(categories);
     });
   });
 });
