+++ frontend/src/queries/templates/useTemplateListQuery.spec.tsx
@@ -3,7 +3,7 @@ import { renderHook, waitFor } from '@testing-library/react';
 
 import { DEFAULT_SORTING_OPTION, PAGE_SIZE } from '@/api';
 import { AuthProvider } from '@/contexts';
-import { templates as mockTemplates } from '@/mocks/templateList.json';
+import { templates as mockTemplates } from '@/mocks/fixtures/templateList.json';
 
 import { useTemplateListQuery } from './useTemplateListQuery';
